Ensure that client messages are handled in order
Due to how we simulate delay, messages from client can be
handled in a different order as they are generated, causing,
for instance, jittery paddle.  By adding a timestamp to
client messages and checking against that before processing
a message from the client, we ensure that messages are
processed in the order that they are generated.

Messages that are out-of-order will be discarded.  Such discards
should not cause much issues since clients send updates
continuously.